First time deploying to robot
I had issues where my tankDriveVelocities method didnt go in the same directions as acrade drive, right was inverted or not inverted between them, not sure why. This still needs to be fixed.

Im getting a weird talon issue where its flashing between a command its given and idle, ie. orange and blinking red/green. It flashed really fast and makes the robot jitter it does when its low battery, but it does it in different situations than when its low on battery.